pyproject.toml
[tool.poetry]
name = "py-bife"
version = "0.1.0"
description = ""
authors = ["Juan Knebel <juanknebel@gmail.com>"]
license = "LGPL"
readme = "README.md"
packages = [{include = "py_bife"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.1"}
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.21"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "run-local:start"
